+++ sdk/storage/storage-blob/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 12.33.1 (Unreleased)
+## 12.34.0 (Unreleased)
 
 ### Features Added